Remove `func_8008627C` from `functions.h`

+++ src/melee/cm/camera.h
@@ -25,7 +25,7 @@ unk_t func_80030E44();
 CameraBox* func_80029020();
 void func_800290D4();
 void func_8002A278(f32 x, f32 y);
-void func_8002E6FC(u8);
+void func_8002E6FC(u32);
 void func_8002E948(u32 (*)(Vec*));
 void func_8002EB5C(f32);
 void func_8002EC7C(f32);
